deletes sidenav2 component and updates imports
